Turno

+++ src/Brasa/TurnoBundle/Repository/TurServicioDetalleRepository.php
@@ -18,7 +18,7 @@ class TurServicioDetalleRepository extends EntityRepository {
     public function pendientesCliente($codigoCliente) {
         $em = $this->getEntityManager();
         $dql   = "SELECT sd FROM BrasaTurnoBundle:TurServicioDetalle sd JOIN sd.servicioRel s "
-                . "WHERE s.codigoClienteFk = " . $codigoCliente;
+                . "WHERE s.codigoClienteFk = " . $codigoCliente . " AND s.estadoCerrado = 0";
         $query = $em->createQuery($dql);
         $arResultado = $query->getResult();
         return $arResultado;                
@@ -27,9 +27,19 @@ class TurServicioDetalleRepository extends EntityRepository {
     public function eliminarSeleccionados($arrSeleccionados) {        
         if(count($arrSeleccionados) > 0) {
             $em = $this->getEntityManager();
-            foreach ($arrSeleccionados AS $codigo) {                
-                $arServicioDetalle = $em->getRepository('BrasaTurnoBundle:TurServicioDetalle')->find($codigo);                
-                $em->remove($arServicioDetalle);                  
+            foreach ($arrSeleccionados AS $codigo) {
+                $intNumeroRegistros = 0;
+                $dql   = "SELECT COUNT(pd.codigoPedidoDetallePk) as numeroRegistros FROM BrasaTurnoBundle:TurPedidoDetalle pd "
+                        . "WHERE pd.codigoServicioDetalleFk = " . $codigo;
+                $query = $em->createQuery($dql);
+                $arrPedidoDetalles = $query->getSingleResult(); 
+                if($arrPedidoDetalles) {
+                    $intNumeroRegistros = $arrPedidoDetalles['numeroRegistros'];
+                }
+                if($intNumeroRegistros <= 0) {
+                    $arServicioDetalle = $em->getRepository('BrasaTurnoBundle:TurServicioDetalle')->find($codigo);                
+                    $em->remove($arServicioDetalle);                                      
+                }
             }                                         
             $em->flush();       
         }
